+++ src/main/java/ca/cal/tp2/model/Livre.java
@@ -6,7 +6,7 @@ public class Livre extends Document {
     private String editeur;
     private int nombrePages;
 
-    public Livre(int documentID, String titre, int nombreExemplaires,
+    public Livre(Long documentID, String titre, int nombreExemplaires,
                  String ISBN, String auteur, String editeur, int nombrePages) {
         super(documentID, titre, nombreExemplaires);
         this.ISBN = ISBN;
